--- a/package.json
+++ b/package.json
@@ -18,13 +18,14 @@
   "main": "index.js",
   "repository": "git://github.com/isaacs/node-lru-cache.git",
   "devDependencies": {
+    "benchmark": "^2.1.4",
     "standard": "^5.4.1",
     "tap": "^10.3.3"
   },
   "license": "ISC",
   "dependencies": {
-    "pseudomap": "^1.0.1",
-    "yallist": "^2.0.0"
+    "pseudomap": "^1.0.2",
+    "yallist": "^2.1.2"
   },
   "files": [
     "index.js"