cc @isaacs 

As it turns out, only removal of `priv` was needed.

Benchmarks before the change:

```
set x 504,171 ops/sec ±1.95% (84 runs sampled)
get x 4,050,237 ops/sec ±0.54% (94 runs sampled)
peek x 5,008,537 ops/sec ±0.70% (92 runs sampled)
set x 489,478 ops/sec ±1.67% (85 runs sampled)
```

After:

```
set x 1,336,330 ops/sec ±3.43% (62 runs sampled)
get x 8,227,208 ops/sec ±1.14% (88 runs sampled)
peek x 8,908,491 ops/sec ±1.07% (86 runs sampled)
set with `maxAge` x 1,113,239 ops/sec ±3.65% (68 runs sampled)
```

Fix #111