@isaacs 2017-06-11
Timing tests are notoriously tricky to get right on CI.  I bumped up the slack in the timeouts significantly when in CI mode, so they're slower, but whatever.

Landed on 4.1.1.  Thanks!